# openapi.yaml
openapi: 3.1.0
info:
  title: Guard API — Антиплагиат БГУИР
  version: "1.0.0"
  description: |
    REST API веб-приложения Guard (антиплагиат БГУИР).

    **Аудитории**
    - Браузерный UI (`index.html` / `app.html` / `admin.html`) — cookie `guard_session`.
    - Плагин Moodle СЭО — заголовок `X-API-Key` (= `MOODLE_API_KEY`, ≥16 символов)
      или `Authorization: Bearer`. Логин студента: поле `username`, query `username`
      или заголовок `X-Moodle-Username`.
    - QR-справки — HMAC-подпись `sig` (`REPORT_ACCESS_SECRET`).
    - Ночной cron — `X-Cron-Secret` (= `CLEANUP_CRON_SECRET`).

    **Эндпоинты плагина Moodle:** `POST /api/upload`, `GET /api/documents/{documentId}/analysis`,
    `GET /api/report/{documentId}/links`.

    **Особенности**
    - `POST /api/auth/register` и `POST /api/report` в текущей сборке без авторизации.
    - `GET /api/analysis/health` требует секрет только если задан `CLEANUP_CRON_SECRET`.
  contact:
    name: БГУИР
servers:
  - url: /
    description: Тот же хост, что и сайт Guard
...
paths:
...
  /api/documents/{documentId}/analysis:
    get:
      tags: [analysis, moodle]
      summary: Статус анализа документа
      security:
        - cookieAuth: []
        - moodleApiKey: []
        - moodleBearer: []
      parameters:
        - { $ref: "#/components/parameters/documentId" }
        - in: query
          name: username
          schema: { type: string }
          description: LDAP uid (для Moodle)
      responses:
        "200":
          description: Статус
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalysisStatus" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/report/{documentId}/links:
    get:
      tags: [reports, moodle]
      summary: Подписанные URL и QR для справки
      security:
        - cookieAuth: []
        - moodleApiKey: []
        - moodleBearer: []
      parameters:
        - { $ref: "#/components/parameters/documentId" }
        - in: query
          name: username
          schema: { type: string }
      responses:
        "200":
          description: Ссылки
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  verifyUrl: { type: string }
                  originalUrl: { type: string }
                  reportPdfUrl: { type: string }
                  verifyQrImage: { type: string }
                  originalQrImage: { type: string }
...
  /api/admin/document-types/{id}:
    patch:
      tags: [admin, directories]
      summary: Обновить тип работы
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
                name: { type: string }
                description: { type: string }
                isActive: { type: boolean }
      responses:
        "200": { description: type }
        "400": { $ref: "#/components/responses/Error" }
    delete:
      tags: [admin, directories]
      summary: Удалить тип работы
      description: |
        Если есть активные (не archived) работы с этим типом — 400 с текстом и числом.
        Иначе физическое удаление. Архивные документы отвязываются (`documentTypeId` → NULL).
      security: [{ cookieAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Удалён
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  unlinkedArchived: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/admin/statistics:
    get:
      tags: [admin]
      summary: Дашборд статистики
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/startDate" }
        - { $ref: "#/components/parameters/endDate" }
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: minUniqueness
          schema: { type: integer }
        - in: query
          name: maxUniqueness
          schema: { type: integer }
        - in: query
          name: minPlagiarism
          schema: { type: integer }
        - in: query
          name: maxPlagiarism
          schema: { type: integer }
      responses:
        "200": { description: statistics }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: guard_session
      description: HMAC-сессия после POST /api/auth/login
    moodleApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Равен MOODLE_API_KEY (≥16 символов)
    moodleBearer:
      type: http
      scheme: bearer
      description: Тот же секрет, что MOODLE_API_KEY
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
      description: Равен CLEANUP_CRON_SECRET
    qrSignature:
      type: apiKey
      in: query
      name: sig
      description: HMAC подпись REPORT_ACCESS_SECRET
  parameters:
    documentId:
      in: path
      name: documentId
      required: true
      schema: { type: integer }
...
    startDate:
      in: query
      name: startDate
      schema: { type: string, format: date }
    endDate:
      in: query
      name: endDate
      schema: { type: string, format: date }
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
        code: { type: string }
...
    AnalysisStatus:
      type: object
      properties:
        success: { type: boolean }
        documentId: { type: integer }
        jobId: { type: integer }
        jobStatus: { type: string }
        documentStatus: { type: string }
        title: { type: string }
        filename: { type: string }
        category: { type: string }
        wordCount: { type: integer }
        localPlagiarismPercent: { type: number }
        originalityPercent: { type: number }
        plagiarismPercent: { type: number }
        plagiarismPercentMl: { type: number }
        aiPercentMl: { type: number }
        processingTimeMs: { type: integer }
        lastError: { type: string, nullable: true }
  responses:
...
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }